requirements.txt
bidict==0.22.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
DBUtils==3.0.2
h11==0.14.0
pycparser==2.21
PyMySQL==1.0.2
python-engineio==4.3.4
python-socketio==5.7.2
uvicorn==0.20.0
websockets==10.4
